requirements.txt
# Core signal processing
mne==1.6.1
pyedflib==0.1.37
numpy==1.26.4
scipy==1.13.0
pandas==2.2.2

# ML / Training
scikit-learn==1.4.2
xgboost==2.0.3
torch==2.3.0
torchvision==0.18.0

# MLflow
mlflow==2.13.0
boto3==1.34.101

# DVC
dvc==3.51.0
dvc-s3==3.2.0

# Django & API
Django==4.2.13
djangorestframework==3.15.1
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
redis==5.0.4
django-redis==5.4.0
gunicorn==22.0.0
whitenoise==6.7.0
dj-database-url==2.1.0

# IoT / MQTT
paho-mqtt==2.0.0

# Monitoring
evidently==0.4.30
prefect==2.19.7

# Utilities
python-dotenv==1.0.1
click==8.1.7
loguru==0.7.2
